update layout building preview for view controller
